Move MenuBar, delete View, delete ViewApi
